ErrorLayer: Use cached display panels instead of recreating them
This is a significant improvement in terms of memory allocations.
The ErrorLayer#paint calls go from 36% of all to 2.6% of all.
This is a 92% (!!!) improvement during map panning.

Due to the improvement in memory allocations, this also decreases
CPU usage by garbage collection from 35% to 14%. This is a 60%
improvement during map panning.

Signed-off-by: Taylor Smock <tsmock@fb.com>