ErrorLayer: Further decrease memory allocations
This brings us from 2.6% of all allocations in ErrorLayer#paint down to
0.78% of all allocations.

This was done by caching the icons for Osmose (2.36% of all) and
Keepright (0.02% of all).

Overall, with "ErrorLayer: Use cached display panels instead of
recreating them" (see preceeding commit), this is a 97% reduction in
memory allocations.

When compared to the previous commit, there was a slight (14% -> 15%)
change in time spent for GC.

Signed-off-by: Taylor Smock <tsmock@fb.com>